Some of Steve's patched fixed, other changes... See following excerpt form the CHANGES file.  Prepping for bigger changes ahead...
	1. In the voxware server, a call to disableProcessFlow() was added
to the end of createServerComponents().  This call effectivly disables the
sig handler that calls intervalProc() on the first client connect after a
server start.  Kill the client and all subsequent ones work fine.  I've
commented out this call.  Anybody know what was supposed to be
accomplished here?  Seems to work (again) fine without it.

	2. In the dia/main.c file, calls were added to close() 0,1, and 2
with the comment that 'we are running in daemon mode so we should close
these'.  Well, au does not run in daemon mode, therefore any debugging
calls (fprintf(stderr,...) are lost later on in the code.  I've commented
these out until (or if) we want to do a 'real' daemon mode.  So now the
question: Should au kick itself in the background by default, and
therefore run as a true daemon?  A cmdline option could be added to do the
current behavior (which is handy for debugging)... What do you think?

	3. I got real tired of the PRMSG macros all over the place, and
#def'd in several files.  I've added a new varargs function to the os/
component called osLogMsg(...).  This is a way better way to handle this
kind of stuff. I've changed the voxware (and sgi) servers to use this
function instead (the only ones using PRMSG). If your working on a server
and want to use these, #include "aulog.h" (in 1.3c+) and go for it.

	4. I've added a NasConfig.h file to config/ that defines (or will)
a few things (like whether you want to use syslog, etc).  Initially you
will need to tweak this manually for your site... This file will be a
prime candidate for generation by autoconf in a future release. No
question here, just a comment.

	5. I've made some changes to the way various include files are
generated during a build... Basically any module that contains header
files needed by other modules have the appropriate header files exported
into TOPLEV/include during the includes phase of the build using the imake
BuildIncludes() macros.  These were removed in the mondo patch - I've
re-added them where appropriate.  Any reason these were removed?