1.4.2a  4/21/2001
	- nasd will now look at AUDIOSERVER and DISPLAY to determine
the correct port to listen on, unless overriden on the command line.
Useful for sites that use the STARTSERVER build option so that the
spawned nasd can listen on a port other than :0


	4/29/2001

	- patch from Steve McIntyre to resolve man page issues, and
add a libm dependancy to libaudio for linux platforms.

	- patches from Philippe Jouguet

		new options in nasd.conf for the voxware server:
		- add 'mixer' device option to secify the mixer.
		- add 'forcerate' option to force the sample rate
			to the current rate
		- add 'autoOpen' option to specify whether the audio
			devices are opened at init time.
		- add 'readwrite' option to specify whether devices
			are opened read/write (default) or
			readonly/writeonly for the input and output
			devices, respectivly.


	- added patch from Nick Ing-Simmons to pass the correct port
offset when starting to nasd for libaudio (if START_SERVER is defined
at build time).

	- always output verbose/debugging info to stderr if debugging
is enabled (even if you normally use syslog).  stderr is not closed
when debugging is turned on.