Narrow a race in auvoxware.c::disableProcessFlow()
The race was found by Paul Fox, reporting it thus to the mailing list:

"""
i believe the problem is that if we take an alarm signal while still
in the middle of disableProcessFlow(), we'll call disableProcessFlow()
again.  so rather than clearing the processFlowEnabled flag when we're
finished disabling, we should probably clear it beforehand, as in this
patch.
"""
  -- http://radscan.com/pipermail/nas/2013-July/001259.html

I modified his proposed patch to not change behaviour on SCO
(but SCO code cannot betested by me).

The Linux code compiles and works for me, but I did not experience this
problem myself.

Stack trace of the problem from Paul Fox was sent to the mailing list as well:
http://radscan.com/pipermail/nas/2013-July/001258.html

Relevant part showing just the race:

 #15 0x08055841 in disableProcessFlow () at auvoxware.c:1342
 #16 0x08051def in AuProcessData () at auprocess.c:988
 #17 0x08055389 in intervalProc (sig=14) at auvoxware.c:1146
 #18 <signal handler called>
 #19 0xb7f007e0 in malloc () from /lib/libc.so.6
 #20 0xb7ff0c44 in ?? () from /lib/ld-linux.so.2
 #21 0xb7ff4ad9 in ?? () from /lib/ld-linux.so.2
 #22 0xb7d95c0b in ?? () from /lib/libdl.so.2
 #23 0xb7ff0b36 in ?? () from /lib/ld-linux.so.2
 #24 0xb7d9609c in ?? () from /lib/libdl.so.2
 #25 0xb7d95b41 in dlopen () from /lib/libdl.so.2
 #26 0xb7debcc8 in snd_dlopen () from /usr/lib/libasound.so.2
 #27 0xb7debf27 in ?? () from /usr/lib/libasound.so.2
 #28 0xb7e26ce3 in ?? () from /usr/lib/libasound.so.2
 #29 0xb7e2705e in snd_pcm_rate_open () from /usr/lib/libasound.so.2
 #30 0xb7e2bd92 in ?? () from /usr/lib/libasound.so.2
 #31 0xb7e2c410 in ?? () from /usr/lib/libasound.so.2
 #32 0xb7e133b4 in ?? () from /usr/lib/libasound.so.2
 #33 0xb7e0e951 in snd_pcm_hw_params () from /usr/lib/libasound.so.2
 #34 0xb7e8ab65 in ?? () from /usr/lib/libalsatoss.so.0
 #35 0xb7e8b294 in ?? () from /usr/lib/libalsatoss.so.0
 #36 0xb7e8d0c7 in lib_oss_pcm_ioctl () from /usr/lib/libalsatoss.so.0
 #37 0xb7fddb6b in ioctl () from /usr/lib/libaoss.so
 #38 0x08055841 in disableProcessFlow () at auvoxware.c:1342