Accept only possible values for listen port offset of nasd.
Verify that the listen port offset specified as a command line argument
to nasd is a non-negative number that will result in a valid TCP port
number if added to AU_DEFAULT_TCP_PORT (currently 8000).

Specifying a long argument starting with a colon would otherwise result
in buffer overflows later on.

The problem was reported to the nas mailing list  by
Hamid Zamani <me@hamidx9.ir>, together with other vulnerabilities
in NAS 1.9.3:

http://radscan.com/pipermail/nas/2013-August/001270.html

[Adding bounds checks to the string operations is still needed to guarantee
they do not overflow.]