nasd/Minix: Sanity check of TCP_DEVICE environment variable.
On Minix, a standard way to select a specific network interface is using
environment variables. For TCP connections this variable is called
TCP_DEVICE. nasd supports this, but used the environment variable,
if set, without checking its value. Ensure that the default TCP device
is a prefix of the environment variables value, ignore the environment
variable otherwise.

This commit addresses the last open vulnerability reported by
Hamid Zamani <me@hamidx9.ir> to the NAS mailing list:

http://radscan.com/pipermail/nas/2013-August/001270.html

NOTE: This change is not even compile tested, because none of the NAS
      developers has a Minix system.